requirements.txt
requests==2.19.1
python-consul==1.1.0